refactor lib.rs into tally.rs
